The size_t shall continue until morale improves.

+++ src/pyjion/pyjit.h
@@ -64,7 +64,7 @@ typedef struct PyjionSettings {
     bool profiling = false;
     unsigned short optimizationLevel = 1;
     int recursionLimit = DEFAULT_RECURSION_LIMIT;
-    int codeObjectSizeLimit = DEFAULT_CODEOBJECT_SIZE_LIMIT;
+    size_t codeObjectSizeLimit = DEFAULT_CODEOBJECT_SIZE_LIMIT;
 
     // Optimizations
     bool opt_inlineIs = OPTIMIZE_IS; // OPT-1
